refactor: create separate folder for helper functions
